\docType{package}
\name{DUPlot}
\alias{DUPlot}
\title{Create high impact visualizations for direct selling industry}
\usage{
DUPlot()
}
\description{
This package gather several visualization designs which representation lets make quick analysis and consequently take better decisions in business.
}
\author{
\itemize{
  \item Erickson Sandoval \email{erickson.sandoval@utec.edu.pe}
  \item Pavel Cotacallapa \email{pcotacallapa@pucp.edu.pe}
}
}
