\name{DUPlot}
\alias{DUPlot}
\title{Create high impact visualizations for direct selling companies}
\usage{
DUPlot()
}
\description{
This package gather several visualization designs which representation lets make quick analysis and consequently take better decisions in business.
}
\examples{
DUPlot()
}
